optmized for sum

+++ src_cuda/fp_exponent.cu
@@ -10,7 +10,7 @@ struct abs_max {
 };
 
 template <class real_t, class real_const_ptr, int32_t BLOCK_THREADS>
-__global__ void vector_exponent_kernel(int32_t M, real_const_ptr A, int64_t lda, int32_t umax, int32_t* __restrict__ vec_expon) {
+__global__ void vector_exponent_kernel(int32_t M, real_const_ptr A, int64_t lda, int32_t umax, uint64_t* __restrict__ vec_expon) {
   __shared__ typename cub::BlockReduce<double, BLOCK_THREADS>::TempStorage temp_reduce;
   cub::BlockReduce<double, BLOCK_THREADS> block_reduce(temp_reduce);
   abs_max amax_func;
@@ -26,18 +26,18 @@ __global__ void vector_exponent_kernel(int32_t M, real_const_ptr A, int64_t lda,
 
   if (threadIdx.x == 0) {
     int32_t e; frexp(threadB, &e);
-    vec_expon[blockIdx.x] = (umax + e) & 0x7fffffff;
+    vec_expon[blockIdx.x] = uint32_t((umax + e) & 0x7fffffff);
   }
 }
 
 constexpr int32_t block_threads = 512;
 
-void internal::int8::vexp_f64(cudaStream_t stream, int32_t M, int32_t N, const double* A, int32_t lda, int32_t umax, int32_t* vec_expon) {
+void internal::int8::vexp_f64(cudaStream_t stream, int32_t M, int32_t N, const double* A, int32_t lda, int32_t umax, uint64_t* vec_expon) {
   vector_exponent_kernel <double, const double* __restrict__, block_threads>
     <<< N, block_threads, 0, stream >>> (M, A, lda, -umax, vec_expon);
 }
 
-void internal::int8::vexp_f32(cudaStream_t stream, int32_t M, int32_t N, const float* A, int32_t lda, int32_t umax, int32_t* vec_expon) {
+void internal::int8::vexp_f32(cudaStream_t stream, int32_t M, int32_t N, const float* A, int32_t lda, int32_t umax, uint64_t* vec_expon) {
   vector_exponent_kernel <float, const float* __restrict__, block_threads>
     <<< N, block_threads, 0, stream >>> (M, A, lda, -umax, vec_expon);
 }
